% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{orb}
\alias{orb}
\title{Find pairs of points within fixed cutoff}
\usage{
orb(x, npartitions, dcell)
}
\arguments{
\item{x}{Matrix of positions where a column corrresponds to a point}

\item{npartitions}{Number of partitions to generate}

\item{dcell}{Side-length if square grid cell used to bin points}
}
\value{
List of bounadries
}
\description{
Find pairs of points within fixed cutoff
}
\examples{
x <- matrix(rnorm(5000), nrow=2)
orb(x, npartitions=8, dcell=0.15)
}
